rename wcr `DynamicPageTitle` & `DynamicPageHeader`
