+++ packages/main/src/components/ObjectPageHeader/ObjectPageHeader.module.css
@@ -4,6 +4,6 @@
   z-index: 1;
   padding-block-start: 1rem;
   padding-block-end: 1rem;
-  display: var(--_ui5wcr_DynamicPage_header_display);
+  display: var(--_ui5wcr_ObjectPage_header_display);
   overflow: hidden;
 }
