+++ packages/main/src/index.ts
@@ -4,10 +4,10 @@ export * from './components/ActionSheet/index.js';
 export * from './components/AnalyticalCardHeader/index.js';
 export * from './components/AnalyticalTable/index.js';
 export * from './components/AnalyticalTable/defaults/LoadingComponent/TablePlaceholder.js';
-export { DynamicPageHeader } from './components/DynamicPageHeader/index.js';
-export type { DynamicPageHeaderPropTypes } from './components/DynamicPageHeader/index.js';
-export { DynamicPageTitle } from './components/DynamicPageTitle/index.js';
-export type { DynamicPageTitlePropTypes } from './components/DynamicPageTitle/index.js';
+export { ObjectPageHeader } from './components/ObjectPageHeader/index.js';
+export type { ObjectPageHeaderPropTypes } from './components/ObjectPageHeader/index.js';
+export { ObjectPageTitle } from './components/ObjectPageTitle/index.js';
+export type { ObjectPageTitlePropTypes } from './components/ObjectPageTitle/index.js';
 export * from './components/ExpandableText/index.js';
 export * from './components/FilterBar/index.js';
 export * from './components/FilterGroupItem/index.js';
