+++ packages/main/src/webComponents/DynamicPage/DynamicPage.mdx
@@ -7,7 +7,7 @@ import { DynamicPageHeader } from '../DynamicPageHeader';
 
 <Meta of={ComponentStories} />
 
-<DocsHeader subComponents={['DynamicPageTitle', 'DynamicPageHeader']} />
+<DocsHeader />
 
 ## Example
