+++ packages/main/src/webComponents/index.ts
@@ -26,7 +26,8 @@ export * from './DateRangePicker/index.js';
 export * from './DateTimePicker/index.js';
 export * from './Dialog/index.js';
 export * from './DynamicPage/index.js';
-// TODO: add DynamicPageHeader & DynamicPageTitle
+export * from './DynamicPageHeader/index.js';
+export * from './DynamicPageTitle/index.js';
 export * from './DynamicSideContent/index.js';
 export * from './FileUploader/index.js';
 export * from './FilterItem/index.js';
